drop BaseCiCdGenerator
