+++ generators/ci-cd/generators/circle/generator.ts
@@ -16,21 +16,43 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import BaseApplicationGenerator from '../../../base-simple-application/index.ts';
-import { BaseCiCdGenerator } from '../../support/generator.ts';
+import BaseSimpleApplicationGenerator from '../../../base-simple-application/index.ts';
+import type { Application as CiCdApplication } from '../../types.ts';
 
-export default class CiCdCircleGenerator extends BaseCiCdGenerator {
+export default class CiCdCircleGenerator extends BaseSimpleApplicationGenerator<CiCdApplication> {
   readonly provider = 'circle' as const;
 
   async beforeQueue() {
-    await super.beforeQueue();
+    if (!this.fromBlueprint) {
+      await this.composeWithBlueprints();
+    }
+
+    await this.dependsOnBootstrap('ci-cd');
+    await this.dependsOnJHipster('jhipster:ci-cd:common');
   }
 
   get writing() {
-    return super.writing;
+    return this.asWritingTaskGroup({
+      async writeFiles({ application }) {
+        if (!this.provider) {
+          return;
+        }
+
+        const rootTemplatesPath = this.fetchFromInstalledJHipster('ci-cd/templates');
+        await this.writeFiles({
+          rootTemplatesPath,
+          blocks: [{ templates: ['.circleci/config.yml'] }],
+          context: application,
+        });
+      },
+    });
   }
 
-  get [BaseApplicationGenerator.WRITING]() {
-    return super[BaseApplicationGenerator.WRITING];
+  get [BaseSimpleApplicationGenerator.WRITING]() {
+    return this.delegateTasksToBlueprint(() => this.writing);
+  }
+
+  shouldAskForPrompts() {
+    return true;
   }
 }
